debug kb hook

+++ src/ui/colors.rs
@@ -9,7 +9,7 @@ use windows::Win32::{
 pub fn on_ctlcolor(wparam: WPARAM, _lparam: LPARAM) -> LRESULT {
     unsafe {
         let hdc = HDC(wparam.0 as *mut core::ffi::c_void);
-        SetTextColor(hdc, COLORREF(GetSysColor(COLOR_WINDOWTEXT) as u32));
+        SetTextColor(hdc, COLORREF(GetSysColor(COLOR_WINDOWTEXT)));
         SetBkMode(hdc, TRANSPARENT);
         let brush: HBRUSH = GetSysColorBrush(COLOR_WINDOW);
         LRESULT(brush.0 as isize)
